Update statisticfunctions.cpp
now the S estimator is a fast version, adapted to C++ from Croux, C., Rousseeuw, P.J. (1992). Time-Efficient Algorithms for Two Highly Robust Estimators of Scale. In: Dodge, Y., Whittaker, J. (eds) Computational Statistics. Physica, Heidelberg. https://doi.org/10.1007/978-3-662-26811-7_58. The algorithm is not a copy. It contains a few performance enhancements and adaptions, It is a bit non-trivial to adapt it to c, because c array indices begin with 0, and in this algorithm, one has 7 variables at once who get modified in different ways and become an index function for an array. The algorithm has been tested with many arrays. It seems to work.